♻️ Extract the command line utility functions into a single module in the new subpackage.
